Udapte visual usuarios

+++ src/main/java/com/matiasbadano/challeng/controllers/AdministradorController.java
@@ -124,7 +124,7 @@ public class AdministradorController {
         curso.setNombre(nombre);
 
         Categoria categoria = new Categoria();
-        categoria.setId((long) categoriaId);
+        categoria.setId(categoriaId);
         curso.setCategoria(categoria);
 
         curso.setTurno(turno);
